last

+++ backend/functions/deleteDevices.function.js
@@ -6,31 +6,38 @@ let Device = require('../models/device.model');
 const del = async (serialNumber, UID) => {
     let gateway = new Gateway();
     let device = new Device();
-    let gatewayDevices = [];    
+    let gatewayDevices = [];
     let i;
 
     try {
         gateway = await Gateway.find({ serial: serialNumber }).populate('devices');
     } catch (err) {
-        return err.message || 'Access error';
+        return ['Access error'];
     }
-    gatewayDevices = gateway[0].devices;    
-    for( i=0; i<gatewayDevices.length;i++){
-        if(gatewayDevices[i].uid==UID){
-            break;
+    if (gateway.length == 0) {
+        return ['Gateway not found'];
+    } else {
+        gatewayDevices = gateway[0].devices;
+        for (i = 0; i < gatewayDevices.length; i++) {
+            if (gatewayDevices[i].uid == UID) {
+                break;
+            }
+            if (i == gatewayDevices.length - 1 && gatewayDevices[i].uid != UID) {
+                return ['Device not found'] ;
+            }
         }
-    }    
-    gatewayDevices.splice(i, 1);
+        gatewayDevices.splice(i, 1);
 
 
-    try {
-        Device.findOneAndDelete({ uid: UID });
-    } catch (err) {
-        return err.message || 'Access error';
+        try {
+            Device.findOneAndDelete({ uid: UID });
+        } catch (err) {
+            return ['Access error'];
+        }
+
+
+        return gatewayDevices;
     }
-
-
-    return gatewayDevices;
 }
 
 module.exports = del;
